♻️ refactor models and enums
