+++ pyrb/model/price.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+
+class CurrentPrice(BaseModel):
+    symbol: str
+    price: int
