+++ pyrb/repository/brokerage/base/portfolio.py
@@ -1,14 +1,8 @@
 import abc
 
-from pydantic import BaseModel, NonNegativeFloat, PositiveFloat, PositiveInt
+from pydantic import NonNegativeFloat
 
-
-class Position(BaseModel):
-    symbol: str  # 종목코드
-    quantity: PositiveInt  # 보유수량
-    sellable_quantity: PositiveInt  # 매도가능수량
-    average_buy_price: PositiveFloat  # 매입단가
-    total_amount: PositiveFloat  # 평가금액
+from pyrb.model.position import Position
 
 
 class Portfolio(abc.ABC):
