+++ pyrb/repository/brokerage/ebest/order_manager.py
@@ -1,7 +1,9 @@
 from requests import HTTPError
 
+from pyrb.enums import OrderType
 from pyrb.exceptions import OrderPlacementError
-from pyrb.repository.brokerage.base.order_manager import Order, OrderManager, OrderType
+from pyrb.model.order import Order
+from pyrb.repository.brokerage.base.order_manager import OrderManager
 from pyrb.repository.brokerage.ebest.client import EbestAPIClient
