+++ tests/test_controller.py
@@ -3,8 +3,9 @@ from pytest_mock import MockerFixture
 from typer.testing import CliRunner
 
 from pyrb.controller.cli.main import app
+from pyrb.enums import OrderSide, OrderStatus, OrderType
 from pyrb.exceptions import InsufficientFundsException
-from pyrb.repository.brokerage.base.order_manager import Order, OrderSide, OrderStatus, OrderType
+from pyrb.model.order import Order
 from pyrb.repository.brokerage.context import RebalanceContext
 from pyrb.service.rebalance import Rebalancer
